requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyTelegramBotAPI==4.14.0
aiohttp==3.9.1
sqlalchemy>=2.0.35
psycopg[binary]==3.2.10
python-dotenv==1.0.0
openpyxl==3.1.2
a2wsgi
